refactor(copilot,jobs): extract streaming + jobs storage internals
- src/lib/copilot/streaming.ts: 625→142 LOC.
  Public factories now delegate to streaming-session.ts (createGenericStreamingSession)
  which itself uses streaming-telemetry.ts (StreamContext, pumpEventQueue, terminal recorders).

- src/lib/jobs/storage.ts: 579→334 LOC.
  Types, schema validation, load/save, mutex, cleanup, and mutateJob moved to
  storage-internals.ts so the public jobStorage object stays under the size budget.

Both removed from size-budget baseline (3 production files remaining).
All 1842 tests pass; lint clean.
