+++ .size-budget-baseline.json
@@ -4,8 +4,6 @@
     "src/hooks/use-ai-focus.ts": 541,
 
 
-    "src/lib/copilot/streaming.ts": 625,
-    "src/lib/jobs/storage.ts": 579,
     "src/lib/operations/active-operations.ts": 725
   },
   "test": {
