+++ src/lib/jobs/storage.ts
@@ -1,243 +1,40 @@
 /**
  * Background Jobs Storage
  *
- * File-based storage for background AI jobs using the storage utils.
- * Persists across API route invocations in dev mode.
+ * File-based storage for background AI jobs persisted via the storage utils.
  *
- * ## Concurrency model
+ * @remarks
+ * **Concurrency model.** All mutating operations route through
+ * {@link withJobsMutation}, which holds a process-local async mutex around
+ * the load → mutate → save sequence. Reads always hit disk (no module-level
+ * cache) so the worker process and web process see the latest committed
+ * state without manual invalidation.
  *
- * All mutating operations go through {@link withJobsMutation}, which holds a
- * process-local async mutex around the load → mutate → save sequence. This
- * eliminates intra-process lost-update races. Reads still hit disk on every
- * call (no module-level cache) so the worker process and the web process
- * always see the latest committed state without needing manual invalidation.
- *
- * Cross-process write races (both web and worker writing the same
- * `background-jobs` file) are NOT solved here — that requires the worker to
- * become the sole writer, which is the Phase 2B migration. See
- * `docs/streaming-architecture-plan.md` (session artifact).
- *
- * The {@link withJobsMutation} primitive enforces a no-re-entry invariant:
- * the callback receives the current schema and must return the new schema
- * synchronously. Calling another mutating `jobStorage.*` method from inside
- * the callback would deadlock and is treated as a programmer error.
+ * Cross-process write races (web + worker writing the same `background-jobs`
+ * file) are NOT solved here — that requires the worker to become the sole
+ * writer (Phase 2B of the streaming architecture plan).
  */
 
-import { logger } from '@/lib/logger';
-import type { TracePropagationHeaders } from '@/lib/observability/context-propagation';
-import type { ClientTriggerMetadata } from '@/lib/observability/trigger-metadata';
-import { readStorage, writeStorage, deleteStorage } from '@/lib/storage/utils';
+import { deleteStorage } from '@/lib/storage/utils';
 
-const log = logger.withTag('JobStorage');
+import {
+  type BackgroundJob,
+  type JobErrorCode,
+  type JobsStorageSchema,
+  type JobStatus,
+  STORAGE_KEY,
+  cleanup,
+  drainMutations,
+  jobLog as log,
+  loadJobs,
+  mutateJob,
+  withJobsMutation,
+} from './storage-internals';
 
-type JobStatus = 'pending' | 'running' | 'completed' | 'failed' | 'cancelled';
-
-/**
- * Structured error codes for failed jobs.
- *
- * Surfaced to the polling client so the UI can route credentials-related
- * failures to a re-auth CTA instead of rendering a generic error string.
- */
-export type JobErrorCode =
-  | 'credentials_missing'
-  | 'credentials_refresh_failed'
-  | 'unknown';
-
-interface JobCausalityContext extends TracePropagationHeaders {
-  capturedAt: string;
-  trigger?: ClientTriggerMetadata;
-}
-
-export interface BackgroundJob<T = unknown> {
-  id: string;
-  type: string;
-  /**
-   * Owner of this job. **REQUIRED** on every job since the multi-tenant
-   * hardening — read/list/cancel endpoints filter by this. Populated at
-   * `/api/jobs` POST from {@link requireUserContext}. Older job records
-   * persisted before this field existed are treated as orphaned and are
-   * never returned to any user.
-   */
-  userId: string;
-  targetId?: string;
-  status: JobStatus;
-  causality?: JobCausalityContext;
-  input: Record<string, unknown>;
-  result?: T;
-  error?: string;
-  /** Machine-readable failure classification; set alongside `error`. */
-  errorCode?: JobErrorCode;
-  /**
-   * Short, user-facing label describing what the executor is doing
-   * right now (e.g. "Running tests…"). Updated incrementally during
-   * long-running jobs so the client can narrate progress.
-   */
-  currentStep?: string;
-  createdAt: string;
-  startedAt?: string;
-  completedAt?: string;
-}
-
-// Storage schema for jobs file
-interface JobsStorageSchema {
-  jobs: Record<string, BackgroundJob>;
-  version: number;
-}
-
-const STORAGE_KEY = 'background-jobs';
-const DEFAULT_SCHEMA: JobsStorageSchema = { jobs: {}, version: 1 };
-
-// Cleanup old jobs periodically (keep last 100, remove completed older than 1 hour)
-const MAX_JOBS = 100;
-const MAX_AGE_MS = 60 * 60 * 1000; // 1 hour
-
-// Schema validator
-function validateSchema(data: unknown): data is JobsStorageSchema {
-  if (typeof data !== 'object' || data === null) return false;
-  const obj = data as Record<string, unknown>;
-  if (typeof obj.jobs !== 'object' || obj.jobs === null) return false;
-  if (typeof obj.version !== 'number') return false;
-  return true;
-}
-
-/**
- * Always reads from disk. Cache removed in Phase 1 of the streaming
- * architecture refactor — the cache was the root cause of the
- * cross-process stale-read bug that surfaced as "worker dispatch failed"
- * after the first job.
- */
-async function loadJobs(): Promise<JobsStorageSchema> {
-  try {
-    return await readStorage<JobsStorageSchema>(STORAGE_KEY, DEFAULT_SCHEMA, validateSchema);
-  } catch (err) {
-    log.warn('Failed to load jobs from storage, using default:', err);
-    return { jobs: {}, version: 1 };
-  }
-}
-
-async function saveJobs(data: JobsStorageSchema): Promise<void> {
-  try {
-    await writeStorage(STORAGE_KEY, data);
-  } catch (err) {
-    log.error('Failed to save jobs to storage:', err);
-    throw err;
-  }
-}
-
-/**
- * Process-local async mutex protecting the load → mutate → save sequence.
- * All mutating ops chain through this single promise so concurrent calls in
- * the same process serialise rather than racing.
- */
-let mutationChain: Promise<unknown> = Promise.resolve();
-let inMutation = false;
-
-/**
- * Atomically load → mutate → save the jobs schema under a process-local
- * mutex. The callback receives the current schema, mutates and returns it
- * (or returns a new schema object), plus an optional `result` value to
- * return to the caller.
- *
- * Nested calls (calling another mutating `jobStorage.*` method from inside
- * the callback) are rejected synchronously to surface deadlocks at the
- * source rather than hanging the process.
- */
-async function withJobsMutation<TResult>(
-  fn: (schema: JobsStorageSchema) => { schema: JobsStorageSchema; result: TResult } | Promise<{ schema: JobsStorageSchema; result: TResult }>,
-): Promise<TResult> {
-  const run = async (): Promise<TResult> => {
-    if (inMutation) {
-      throw new Error(
-        'jobStorage: nested mutation detected. Mutating methods must not be called from inside withJobsMutation callbacks.',
-      );
-    }
-    inMutation = true;
-    try {
-      const schema = await loadJobs();
-      const { schema: nextSchema, result } = await fn(schema);
-      await saveJobs(nextSchema);
-      return result;
-    } finally {
-      inMutation = false;
-    }
-  };
-
-  const next = mutationChain.then(run, run);
-  // Keep the chain alive even on rejection so subsequent mutations still run.
-  mutationChain = next.catch(() => undefined);
-  return next;
-}
-
-function cleanup(schema: JobsStorageSchema): JobsStorageSchema {
-  const now = Date.now();
-  const toDelete: string[] = [];
-  const jobs = schema.jobs;
-
-  for (const [id, job] of Object.entries(jobs)) {
-    // Remove old completed/failed/cancelled jobs
-    if (job.status === 'completed' || job.status === 'failed' || job.status === 'cancelled') {
-      const completedAt = job.completedAt ? new Date(job.completedAt).getTime() : 0;
-      if (now - completedAt > MAX_AGE_MS) {
-        toDelete.push(id);
-      }
-    }
-  }
-
-  // If still too many, remove oldest completed first
-  const remainingCount = Object.keys(jobs).length - toDelete.length;
-  if (remainingCount > MAX_JOBS) {
-    const completed = Object.entries(jobs)
-      .filter(([id, job]) =>
-        (job.status === 'completed' || job.status === 'failed' || job.status === 'cancelled') &&
-        !toDelete.includes(id)
-      )
-      .sort((a, b) =>
-        new Date(a[1].createdAt).getTime() - new Date(b[1].createdAt).getTime()
-      );
-
-    const excess = remainingCount - MAX_JOBS;
-    for (let i = 0; i < excess && i < completed.length; i++) {
-      toDelete.push(completed[i][0]);
-    }
-  }
-
-  if (toDelete.length === 0) {
-    return schema;
-  }
-
-  const newJobs = { ...jobs };
-  for (const id of toDelete) {
-    delete newJobs[id];
-  }
-
-  log.debug(`Cleaned up ${toDelete.length} old jobs`);
-  return { ...schema, jobs: newJobs };
-}
-
-/**
- * Internal mutation primitive — applies `mutate(job)` to the job with the
- * given id, persists the result, and returns the updated job. Returns
- * `undefined` if the job does not exist.
- *
- * All public mutating methods are thin wrappers over this.
- */
-async function mutateJob<T>(
-  id: string,
-  mutate: (job: BackgroundJob<T>) => BackgroundJob<T>,
-): Promise<BackgroundJob<T> | undefined> {
-  return withJobsMutation<BackgroundJob<T> | undefined>((schema) => {
-    const job = schema.jobs[id] as BackgroundJob<T> | undefined;
-    if (!job) return { schema, result: undefined };
-    const updated = mutate(job);
-    schema.jobs[id] = updated as BackgroundJob;
-    return { schema, result: updated };
-  });
-}
+export type { BackgroundJob, JobErrorCode, JobStatus };
 
 export const jobStorage = {
-  /**
-   * Create a new job.
-   */
+  /** Create a new job. */
   async create<T>(job: Omit<BackgroundJob<T>, 'status' | 'createdAt'>): Promise<BackgroundJob<T>> {
     if (!job.userId) {
       throw new Error('jobStorage.create: userId is required (multi-tenant invariant)');
@@ -256,22 +53,13 @@ export const jobStorage = {
   },
 
   /**
-   * Atomic check-then-create.
+   * Atomic check-then-create. Closes the read-then-write race that existed
+   * when callers used `get(id)` + `create(...)` separately.
    *
-   * Closes the read-then-write race that existed when callers used
-   * `get(id)` + `create(...)` separately: two concurrent requests
-   * with the same `id` could both observe "no existing record" before
-   * either insert, and the second `create` would silently clobber
-   * the first.
-   *
-   * `findCollision` runs INSIDE the mutation, with the freshly-loaded
-   * schema, so callers can also enforce tuple uniqueness (e.g. the
-   * chat-response `(userId, threadId, assistantMessageId)` dedupe)
-   * under the same lock that the insert holds.
-   *
-   * - If `findCollision` returns a job → `{ created: false, existing }`.
-   * - If `schema.jobs[id]` already exists → `{ created: false, existing }`.
-   * - Otherwise → `{ created: true, job }` with the freshly-inserted record.
+   * `findCollision` runs INSIDE the mutation with the freshly-loaded schema,
+   * so callers can enforce tuple uniqueness (e.g. the chat-response
+   * `(userId, threadId, assistantMessageId)` dedupe) under the same lock
+   * that holds the insert.
    */
   async createIfAbsent<T>(
     job: Omit<BackgroundJob<T>, 'status' | 'createdAt'>,
@@ -280,49 +68,43 @@ export const jobStorage = {
     if (!job.userId) {
       throw new Error('jobStorage.createIfAbsent: userId is required (multi-tenant invariant)');
     }
-    return withJobsMutation<{ created: true; job: BackgroundJob<T> } | { created: false; existing: BackgroundJob<T> }>(
-      (schema) => {
-        const cleaned = cleanup(schema);
-        const byId = cleaned.jobs[job.id];
-        if (byId) {
-          return { schema: cleaned, result: { created: false, existing: byId as BackgroundJob<T> } };
-        }
-        const collision = findCollision ? findCollision(cleaned.jobs) : undefined;
-        if (collision) {
-          return { schema: cleaned, result: { created: false, existing: collision as BackgroundJob<T> } };
-        }
-        const newJob: BackgroundJob<T> = {
-          ...job,
-          status: 'pending',
-          createdAt: new Date().toISOString(),
-        };
-        cleaned.jobs[job.id] = newJob as BackgroundJob;
-        log.info(`Created job (if-absent): ${job.id} (${job.type})`);
-        return { schema: cleaned, result: { created: true, job: newJob } };
-      },
-    );
+    return withJobsMutation<
+      { created: true; job: BackgroundJob<T> } | { created: false; existing: BackgroundJob<T> }
+    >((schema) => {
+      const cleaned = cleanup(schema);
+      const byId = cleaned.jobs[job.id];
+      if (byId) {
+        return { schema: cleaned, result: { created: false, existing: byId as BackgroundJob<T> } };
+      }
+      const collision = findCollision ? findCollision(cleaned.jobs) : undefined;
+      if (collision) {
+        return { schema: cleaned, result: { created: false, existing: collision as BackgroundJob<T> } };
+      }
+      const newJob: BackgroundJob<T> = {
+        ...job,
+        status: 'pending',
+        createdAt: new Date().toISOString(),
+      };
+      cleaned.jobs[job.id] = newJob as BackgroundJob;
+      log.info(`Created job (if-absent): ${job.id} (${job.type})`);
+      return { schema: cleaned, result: { created: true, job: newJob } };
+    });
   },
 
-  /**
-   * Get a job by ID.
-   */
+  /** Get a job by ID. */
   async get<T>(id: string): Promise<BackgroundJob<T> | undefined> {
     const schema = await loadJobs();
     return schema.jobs[id] as BackgroundJob<T> | undefined;
   },
 
-  /**
-   * Update a job's status.
-   */
+  /** Update a job's fields. */
   async update<T>(id: string, updates: Partial<BackgroundJob<T>>): Promise<BackgroundJob<T> | undefined> {
     const updated = await mutateJob<T>(id, (job) => ({ ...job, ...updates } as BackgroundJob<T>));
     if (updated) log.debug(`Updated job ${id}: status=${updated.status}`);
     return updated;
   },
 
-  /**
-   * Mark a job as running.
-   */
+  /** Mark a job as running. */
   async markRunning(id: string): Promise<BackgroundJob | undefined> {
     return mutateJob(id, (job) => ({
       ...job,
@@ -331,9 +113,7 @@ export const jobStorage = {
     }));
   },
 
-  /**
-   * Mark a job as completed with result.
-   */
+  /** Mark a job as completed with a result. */
   async markCompleted<T>(id: string, result: T): Promise<BackgroundJob<T> | undefined> {
     return mutateJob<T>(id, (job) => ({
       ...job,
@@ -344,10 +124,9 @@ export const jobStorage = {
   },
 
   /**
-   * Mark a job as failed with error.
-   *
-   * Optionally accepts a structured `errorCode` so polling clients can
-   * distinguish credentials-expired failures from generic errors.
+   * Mark a job as failed. Optionally accepts a structured `errorCode` so
+   * polling clients can distinguish credentials-expired failures from
+   * generic errors.
    */
   async markFailed(
     id: string,
@@ -365,9 +144,7 @@ export const jobStorage = {
 
   /**
    * Update the human-readable `currentStep` label for an in-flight job.
-   *
-   * Safe to call repeatedly; only persists when the value actually
-   * changes to avoid disk churn during high-frequency narration.
+   * Only persists when the value actually changes to avoid disk churn.
    */
   async setCurrentStep(id: string, step: string): Promise<BackgroundJob | undefined> {
     return mutateJob(id, (job) => {
@@ -376,9 +153,7 @@ export const jobStorage = {
     });
   },
 
-  /**
-   * Mark a job as cancelled.
-   */
+  /** Mark a job as cancelled. */
   async markCancelled(id: string): Promise<BackgroundJob | undefined> {
     log.info(`Marking job ${id} as cancelled`);
     return mutateJob(id, (job) => ({
@@ -390,14 +165,9 @@ export const jobStorage = {
   },
 
   /**
-   * Idempotent completion mark.
-   *
-   * Phase 5 introduced compare-and-swap helpers so the worker's terminal
-   * sequence can race safely with DELETE-initiated cancellations. If the
-   * job is already in a terminal state, this is a no-op and returns that
-   * status. Transient I/O failures are retried up to 3× with exponential
-   * backoff (100/200/400 ms) so a flaky disk doesn't corrupt the
-   * happy-path commit.
+   * Idempotent completion mark with retry on transient I/O. If the job is
+   * already terminal, this is a no-op and returns that status. Transient
+   * failures are retried up to 3× with exponential backoff (100/200/400 ms).
    */
   async markCompletedIdempotent<T>(id: string, result: T): Promise<JobStatus> {
     let lastErr: unknown;
@@ -429,10 +199,9 @@ export const jobStorage = {
   },
 
   /**
-   * Atomic CAS: transition to `failed` only if the job is currently
-   * pending or running. Returns `{ status, transitioned }` so the
-   * worker can detect a concurrent DELETE-initiated cancellation and
-   * preserve the user-intent annotation instead of overwriting it.
+   * Atomic CAS: transition to `failed` only if the job is currently pending
+   * or running. Returns `{ status, transitioned }` so the worker can detect
+   * a concurrent DELETE-initiated cancellation and preserve user intent.
    */
   async markFailedIfNonTerminal(
     id: string,
@@ -458,9 +227,8 @@ export const jobStorage = {
 
   /**
    * Atomic CAS: transition to `cancelled` only if the job is currently
-   * pending or running. Returns `{ status, transitioned }` so the
-   * DELETE handler (and any other caller) can tell whether the
-   * cancellation actually moved state forward.
+   * pending or running. Returns `{ status, transitioned }` so the caller
+   * can tell whether the cancellation actually moved state forward.
    */
   async markCancelledIfNonTerminal(
     id: string,
@@ -481,48 +249,38 @@ export const jobStorage = {
     });
   },
 
-  /**
-   * Get all jobs of a specific type.
-   */
+  /** Get all jobs of a specific type. */
   async getByType(type: string): Promise<BackgroundJob[]> {
     const schema = await loadJobs();
-    return Object.values(schema.jobs).filter(job => job.type === type);
+    return Object.values(schema.jobs).filter((job) => job.type === type);
   },
 
-  /**
-   * Get all pending/running jobs.
-   */
+  /** Get all pending/running jobs. */
   async getActive(): Promise<BackgroundJob[]> {
     const schema = await loadJobs();
     return Object.values(schema.jobs).filter(
-      job => job.status === 'pending' || job.status === 'running'
+      (job) => job.status === 'pending' || job.status === 'running',
     );
   },
 
-  /**
-   * Find the active chat-response job for a thread.
-   */
+  /** Find the active chat-response job for a thread. */
   async getActiveChatJobForThread(threadId: string): Promise<BackgroundJob | undefined> {
     const schema = await loadJobs();
     return Object.values(schema.jobs).find(
-      job =>
+      (job) =>
         job.type === 'chat-response' &&
         job.targetId === threadId &&
-        (job.status === 'pending' || job.status === 'running')
+        (job.status === 'pending' || job.status === 'running'),
     );
   },
 
-  /**
-   * Get all jobs (for debugging).
-   */
+  /** Get all jobs (debugging). */
   async getAll(): Promise<BackgroundJob[]> {
     const schema = await loadJobs();
     return Object.values(schema.jobs);
   },
 
-  /**
-   * Delete a job.
-   */
+  /** Delete a job by id. */
   async delete(id: string): Promise<boolean> {
     return withJobsMutation<boolean>((schema) => {
       if (!schema.jobs[id]) return { schema, result: false };
@@ -532,10 +290,8 @@ export const jobStorage = {
   },
 
   /**
-   * Delete every job owned by the given user. Returns the number of
-   * jobs removed. Used by the per-user "delete all my data" endpoint.
-   * Does NOT cancel running jobs — callers should call `cancelRunningJob`
-   * for any jobs that may still be executing in-process.
+   * Delete every job owned by `userId`. Returns the removed ids. Used by the
+   * per-user "delete all my data" endpoint. Does NOT cancel running jobs.
    */
   async deleteForUser(userId: string): Promise<{ deleted: number; ids: string[] }> {
     if (!userId) {
@@ -557,23 +313,22 @@ export const jobStorage = {
     });
   },
 
-  /**
-   * Clear all jobs (for testing).
-   */
+  /** Clear all jobs (testing). Drains pending mutations first. */
   async clear(): Promise<void> {
-    // Drain any pending mutations before deleting the file so we don't race
-    // with an in-flight write.
-    await mutationChain.catch(() => undefined);
+    await drainMutations();
     await deleteStorage(STORAGE_KEY);
   },
 
   /**
-   * @deprecated The module-level cache was removed in the Phase 1 storage
-   * refactor; reads always hit disk now. This method is a no-op kept for
-   * call-site compatibility and will be removed once all defensive
-   * `invalidateCache()` calls have been cleaned up.
+   * @deprecated The module-level cache was removed in Phase 1 of the storage
+   * refactor; reads always hit disk. Kept as a no-op for call-site
+   * compatibility until defensive `invalidateCache()` calls are cleaned up.
    */
   invalidateCache(): void {
     // intentionally empty
   },
 };
+
+// Re-export the schema type for tests/internal consumers that need to mock
+// or inspect on-disk shape directly.
+export type { JobsStorageSchema };
